requirements.txt
numpy
pandas
torch
tushare
pymysql
sqlalchemy
qlib
d
einops==0.8.1
huggingface_hub==0.33.1
matplotlib==3.9.3
pandas==2.2.2
tqdm==4.67.1
safetensors==0.6.2

python-binance